requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
